Added Response Validator
If Response is successfull than response code is 200 (OK) or if response failed than code is 500 (internal server error). after submitting json response we get back result in the form of json with - 'msg', 'response code' & 'status' keys according to response code we get. ie according to if it get success or failed.